+++ src/main/java/net/tassia/pancake/PancakeLauncher.kt
@@ -4,7 +4,7 @@ import net.tassia.pancake.util.config.ConfigIO
 import net.tassia.pancake.util.config.driver.ConfigIniDriver
 import net.tassia.pancake.util.DatabaseConnector
 import net.tassia.pancake.io.PancakeConfig
-import net.tassia.pancake.io.PancakeIO
+import net.tassia.pancake.util.PancakeIO
 import java.io.File
 
 /**
