+++ src/main/java/net/tassia/pancake/util/Readers.kt
@@ -1,4 +1,4 @@
-package net.tassia.pancake.io
+package net.tassia.pancake.util
 
 import java.io.EOFException
 import java.io.Reader
